- stream compressors: use Intel SSE4.2 CRC32C instruction when crc32c is available
- StreamDecompressor: faster decompression via a lot less memcpy