@martindurant 2021-11-18
I am not able to review the changes here - anyone else around?
@paulwouters , it may be enough to see that all tests pass, but would appreciate posting some benchmarks.

Do you know is snappy in [cramjam](https://github.com/milesgranger/pyrus-cramjam) can benefit from similar changes? (cc @milesgranger )