@paulwouters 2021-11-18
For us, the difference is that without these two patches, decompressing ~50mb takes 13sec, ~100mb takes 60sec (this is somewhat data-dependent).

With the patches applied over 0.6.0, decompressing that same 100mb takes 1sec