@martindurant 2021-11-18
Wow! Such a difference doesn't entirely seem plausible :| 
Is this perhaps data where no real compression happened, i.e., what was the compression ratio?