@paulwouters 2021-11-26
These are the benchmarks we are seeing with and without the patch using snappy's own benchmarking tools:

without patch:
-------------------------------------------------------------------------------------------------- benchmark: 4 tests -------------------------------------------------------------------------------------------------
Name (time in ms)                                   Min                   Max                  Mean            StdDev                Median               IQR            Outliers     OPS            Rounds  Iterations
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_snappy_framed[100MB_random-cramjam]       139.5540 (1.0)        149.4682 (1.0)        144.3223 (1.0)      3.8449 (1.45)       145.1978 (1.0)      6.5880 (1.84)          2;0  6.9289 (1.0)           7           1
test_snappy_framed[100MB_random-snappy]      1,037.5058 (7.43)     1,051.2743 (7.03)     1,045.1163 (7.24)     5.1236 (1.93)     1,045.3530 (7.20)     6.5106 (1.82)          2;0  0.9568 (0.14)          5           1
test_snappy_framed[100MB_zeroes-cramjam]       141.6173 (1.01)       149.6211 (1.00)       145.6500 (1.01)     2.6524 (1.0)        145.7977 (1.00)     3.5852 (1.0)           3;0  6.8658 (0.99)          8           1
test_snappy_framed[100MB_zeroes-snappy]      1,046.7271 (7.50)     1,057.7755 (7.08)     1,051.4539 (7.29)     4.4143 (1.66)     1,051.8373 (7.24)     6.6645 (1.86)          2;0  0.9511 (0.14)          5           1
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

With patch:

----------------------------------------------------------------------------------------------- benchmark: 4 tests ----------------------------------------------------------------------------------------------
Name (time in ms)                                 Min                 Max                Mean             StdDev              Median                IQR            Outliers     OPS            Rounds  Iterations
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_snappy_framed[100MB_random-cramjam]     132.7121 (1.0)      145.2343 (1.0)      141.2591 (1.0)       4.3443 (1.04)     143.4440 (1.00)      4.4397 (1.0)           1;1  7.0792 (1.0)           7           1
test_snappy_framed[100MB_random-snappy]      149.3139 (1.13)     177.8449 (1.22)     164.5416 (1.16)     10.4677 (2.52)     166.1130 (1.16)     16.8984 (3.81)          2;0  6.0775 (0.86)          7           1
test_snappy_framed[100MB_zeroes-cramjam]     138.2131 (1.04)     150.1087 (1.03)     143.1957 (1.01)      4.1573 (1.0)      143.0023 (1.0)       5.9779 (1.35)          2;0  6.9834 (0.99)          7           1
test_snappy_framed[100MB_zeroes-snappy]      162.4972 (1.22)     216.4342 (1.49)     173.9976 (1.23)     18.9340 (4.55)     168.5015 (1.18)      5.4640 (1.23)          1;1  5.7472 (0.81)          7           1
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

That is a significant difference.

We would really prefer not to have to maintain a separate patch indefinitely. But the performance increase is essential for our use.